#35055 DTXCをロングノーツ終了位置の配置に対応。(実験で使う譜面への配置用です。DTXM本体は対応していませんのでご注意ください。) #35055 ロングノーツ対応の準備。CChipクラスにノーツの終了位置の項目を追加。 #30806 ボーナスチップの構造を改良。詳細はチケットにて。
#xxxxx SlimDXのソースを最新のdtxmaniaのものに入れ替え。ただし特に意味は無い。

+++ trunk/SlimDXc_Jun2010(VC++2008)/source/direct3d11/StreamOutputStatistics11.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 /*
-* Copyright (c) 2007-2012 SlimDX Group
+* Copyright (c) 2007-2010 SlimDX Group
 * 
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -31,8 +31,8 @@ namespace Direct3D11
 {
 	StreamOutputStatistics::StreamOutputStatistics( const D3D11_QUERY_DATA_SO_STATISTICS& native )
 	{
-		PrimitivesWritten = native.NumPrimitivesWritten;
-		StorageNeeded = native.PrimitivesStorageNeeded;
+		PrimitivesWritten = static_cast<long>( native.NumPrimitivesWritten );
+		StorageNeeded = static_cast<long>( native.PrimitivesStorageNeeded );
 	}
 	
 	D3D11_QUERY_DATA_SO_STATISTICS StreamOutputStatistics::CreateNativeVersion()
